Move ArgparseCliParser to own module
